Change floats to doubles and use fmod from math.h for double remainders.

+++ sistineChapel/robotArm.h
@@ -9,19 +9,19 @@ class RobotArm
 	private:
 		// deadZoneAngle used for deciding how close the joints need to be to 
 		// their target angle to be considered in the correct position
-		float deadZoneAngle;
+		double deadZoneAngle;
 
 		// Elbow parameters
 		int elbowMotorNumber;
 		int elbowEncoderNumber;
-		float elbowTargetAngle;
+		double elbowTargetAngle;
 
 		// Instances of the MotorController and Multiplexer
 		MotorController motor;
 		Multiplexer mux;
 
 		// Contains the logic to actually move a joint to the desired angle
-		bool moveMotorToEncoderAngle(int motorNumber, int encoderNumber, float angle);
+		bool moveMotorToEncoderAngle(int motorNumber, int encoderNumber, double angle);
 	public:	
 		// Constructor
 		RobotArm();
@@ -33,7 +33,7 @@ class RobotArm
 		void setup();
 
 		// Move the elbow the the provided angle
-		void elbow(float degrees);
+		void elbow(double degrees);
 
 		// Run in arduino loop()
 		void loop();
